+++ examples/rowsort.rs
@@ -18,7 +18,8 @@ impl sqllogictest::DB for FakeDB {
 
     fn run(&mut self, sql: &str) -> Result<String, FakeDBError> {
         if sql == "select * from example_rowsort" {
-            // Even if the order is not the same as `slt` file, sqllogictest will sort them before comparing.
+            // Even if the order is not the same as `slt` file, sqllogictest will sort them before
+            // comparing.
             return Ok("1 10 2333\n2 20 2333\n10 100 2333".into());
         }
         unimplemented!("unsupported SQL: {}", sql);